No more Required Reviewers section in Pull Request body (#84)
## Summary:

We've gotten rid of the Required Reviewers section in the Pull Request body because having it in the body means that it needs to be updated every time you push to a pull request. This is fine and doable, but the problem is that if you are editing a pull request on the GitHub web interface, and this action updates the pull request on the GitHub backend, GitHub provides no lock to ensure that both of your changes and the bot's changes can go through... instead, when you hit "Update Comment" on GitHub, it will simply error-out and delete your changes, which is annoying. Now, the required reviewers section will be added back after you land!

Issue: WEB-2589

## Test plan:
Add a Github username (not your own) to the REVIEWERS file and use an exclamation point at the end of your name to denote yourself as a required reviewer.
Commit, push, and come back to this pull request to see that the name you added is now requested as a reviewer, and in a comment that lists out all required reviewers.
Confirm that there is no Required Reviewers section in the pull request body.
Try landing the pull request without approval and see that it still functions as a required reviewer.

Author: yipstanley

Reviewers: lillialexis

Required Reviewers: 

Approved by: lillialexis

Checks: 

Pull request URL: https://github.com/Khan/our-lovely-cli/pull/84