Added the equivalent of "Not Differential Revision" to CODEOWNERS code (#85)
## Summary:

In Herald, there's a condition that allows people to make rules that notify them when someone lands a diff that doesn\'t go through a differential revision. To mimick this functionality in Github, there\'s an additional section in `.github/NOTIFIED` for people to add rules that will update them when someone pushes to `master` or `develop`. If the files changed or the diff matches the rule that they used in `.github/NOTIFIED`, it will tag the usernames in the commit.

Note: the comments on commits are made by the person pushing the code. This is because Github-actions[bot] doesn't have the permissions to make comments on commits.

Issue: WEB-2492

## Test plan:
Either add WEB-2492 to the list of branches in `.github/workflows/pr-notify-push.yml` or checkout `develop` or `master`.
Add yourself to the bottom of the `.github/NOTIFIED` file (below the 'ON PUSH WITHOUT PULL REQUEST' header. Make sure that the condition will match `.github/NOTIFIED`.
Commit and push the change.
Navigate to the commit page (github.com/Khan/our-lovely-cli/commit/<sha>).
Verify that you've made a comment that tags yourself.

Author: yipstanley

Reviewers: lillialexis, yipstanley

Required Reviewers: 

Approved by: lillialexis

Checks: ✅ Flow Coverage, ✅ Jest Coverage, ✅ Jest, ✅ Eslint, ✅ notify_and_request, ✅ autofix, ✅ lint_and_unit

Pull request URL: https://github.com/Khan/our-lovely-cli/pull/85