## Summary:
In #19, we updated the  function in  to ignore merge commits. This involved changing the functionality from doing a squashed comparison between the before and after of a push to doing a comparison between each commit, ignoring merge commits. In #19, @jeresig requested tests. Because adding tests would involve making a rather large diff, the tests were made here, in a separate branch. These tests mock many of the functions used in , including internal functions such as  and external functions such as .

Issue: WEB-2655

## Test plan:
This is part of the test plan for #19