Sometimes commit.verification was undefined, causing Gerald to crash (#41)
## Summary:

This is a small change to Gerald that ensures that Gerald doesn't crash if a commit doesn't have a verification object. Sometimes, commits don't have a verification object, and Gerald will get a property of the verification object, which would crash if the verification object doesn't exist. This PR changes Gerald so that we also check if the verification object exists before †rying to get that property. If the verification object doesn't exist, we treat it as if the commit is not verified.

## Test plan:
I found this bug while making commits to the `debug` branch. Previously, Gerald would straight-up crash, and it doesn't do that when I push to `debug` anymore.

Author: yipstanley

Reviewers: jeresig

Required Reviewers: 

Approved by: jeresig

Checks: ✅ Flow Coverage, ✅ Jest Coverage, ✅ Jest, ✅ Eslint, ✅ lint_and_unit, ✅ autofix, ✅ build_index

Pull request URL: https://github.com/Khan/gerald/pull/41