Match regex rules against the squashed diff in runOnPush (#40)
## Summary:

Previously, Gerald would test regex rules against the diff between each commit in a push and ignore merge commits and validated commits. This was fine, except it would not work very well for the case where someone made a commit that matched a rule, and then reverted that commit or change in the next commit. This PR changes Gerald so that we look at the squashed diff *and* the commit's diff so that we only notify people on a commit if that commit includes changes that match their regex *and* actually ned up in the final code.

Issue: WEB-2762

## Test plan:
Notice how this commit doesn't notify @yipstanley for changes to `src/runOnPush.js`, even though the `.github/NOTIFIED` file has a rule that notifies @yipstanley on changes that match `/\+/m`, which the change to `src/runOnPush.js` satisfies.
https://github.com/Khan/gerald/commit/60bf3c4f32be44831778a6917df77c10546f02f5

That's because this commit was pushed along with this commit, which removed the line that was added in `src/runOnPush.js`. This means that the final squashed diff doesn't actually change `src/runOnPush.js`, so Gerald won't consider that an actual change, whereas the change in `src/runOnComment.js` didn't get undone, so it is counted as an actual change.
https://github.com/Khan/gerald/commit/f68fbeb98daf4d1d3f19c5f860996834dd969a22

Author: yipstanley

Reviewers: jeresig

Required Reviewers: 

Approved by: jeresig

Checks: ✅ Flow Coverage, ✅ Jest Coverage, ✅ Jest, ✅ Eslint, ✅ lint_and_unit, ✅ build_index, ✅ autofix

Pull request URL: https://github.com/Khan/gerald/pull/40