Added a way of running Regex patterns against the content of files changed (#51)
## Summary:

In the Gerald CoCoW, we talked about having a way of running Regular Expressions against the `git diff` of a file, as well as against the entire contents of a file that had been modified. The big difference here is that sometimes, content may be changed in a function, for example, that significantly changes the function's logic, but the function name may not show up in the `git diff`, and if someone had a rule that matched the function name, then the rule wouldn't work, even though this seems like a perfectly reasonable use case for the Regex matching. This is now possible by suffixing a rule with a flag `--match-contents`. Gerald will interpret that flag as the desire to run the Regular Expression against all changed files as opposed to just against the result of `git diff`.

Issue: WEB-2792

## Test plan:
Added unit tests
https://github.com/Khan/gerald/pull/50 is a PR that matches a rule using the `--match-contents` flag.
https://github.com/Khan/gerald/commit/17a3d49569942effcaf4016cc6192eb442d61ba8 is a pushed commit that matches a rule using the `match-contents` flag.

Author: yipstanley

Reviewers: jeresig, yipstanley

Required Reviewers: 

Approved by: jeresig

Checks: ✅ Flow Coverage, ✅ Jest Coverage, ✅ Jest, ✅ Eslint, ✅ autofix, ✅ build_index, ✅ lint_and_unit, ✅ gerald

Pull request URL: https://github.com/Khan/gerald/pull/51