Add label support to REVIEWERS and NOTIFIED files.
This lets you specify rules like this:
```
allfiles: **/* @notifyme
```

And when we add the NOTIFIED section to the github PR, it will now say:
```
@notifyme (allfiles): file1 file2 file3
```
(For rules without labels, the output is unchanged.)

This is helpful for knowing _why_ you were being notified for a
particular PR, since as we add more NOTIFIED rules it can get harder
to tell.

I added the support for REVIEWERS as well, since it was easy, though
I'm not sure how useful it is there.

While in the area, I couldn't help but clean up the existing code a
bit, especially simplifying some of the regexp-parsing logic (at the
cost of more parentheses in the regexp).

Issue: https://khanacademy.atlassian.net/browse/FEI-5970

Test plan:
yarn flow
yarn jest