Add labels in a more principled way.
Before, I was just munging the label to the name, so I wouldn't have
to change a lot of code.  This didn't work because the name was used
in places where the code needs it to actually be a name (e.g. when
telling github what reviewers to add to the PR).

So now I store names and labels separately, and just print the label
explicitly when emitting the comment text.  This also gives me more
control over where the label goes; all in all a more maintainable
arrangement.

Issue: https://khanacademy.atlassian.net/browse/FEI-5970

Test plan:
yarn jest